fix(biscuit): firmware sources compile clean under -Wall -Wextra
uint32_t is unsigned long on Xtensa, so print it with %lu and a cast; brace
the one-line bodies gcc flags as misleading indentation.

+++ apps/biscuit/firmware/src/board.cpp
@@ -194,7 +194,7 @@ int monthDays(int year, int month) {
 
 bool boardBegin() {
   if (panel && touchReady) return true;
-  Serial.printf("Board: PSRAM=%u bytes\n", ESP.getPsramSize());
+  Serial.printf("Board: PSRAM=%lu bytes\n", (unsigned long)ESP.getPsramSize());
   if (!psramFound()) { Serial.println("Board: OPI PSRAM is required"); return false; }
   pinMode(kBacklight, OUTPUT);
   digitalWrite(kBacklight, LOW);
